feat: use unified MakeErrorResponse function
